Separated Preferences tests and PreferencesRaytracing in separate files
